+++ src/capturer/shared.rs
@@ -30,8 +30,6 @@ pub struct SharedCapturer<'a> {
   ctx: &'a DuplicationContext,
   texture: ID3D11Texture2D,
   texture_desc: D3D11_TEXTURE2D_DESC,
-  last_pointer_shape_buffer: Vec<u8>,
-  last_pointer_shape_buffer_size: usize,
   pointer_shape_buffer: Vec<u8>,
   pointer_shape_buffer_size: usize,
 }
@@ -46,8 +44,6 @@ impl<'a> SharedCapturer<'a> {
       texture,
       texture_desc,
       ctx,
-      last_pointer_shape_buffer: Vec::new(),
-      last_pointer_shape_buffer_size: 0,
       pointer_shape_buffer: Vec::new(),
       pointer_shape_buffer_size: 0,
     })
@@ -62,8 +58,6 @@ impl<'a> SharedCapturer<'a> {
       texture,
       texture_desc,
       ctx,
-      last_pointer_shape_buffer: Vec::new(),
-      last_pointer_shape_buffer_size: 0,
       pointer_shape_buffer: Vec::new(),
       pointer_shape_buffer_size: 0,
     })
@@ -182,7 +176,6 @@ impl<'a> Capturer for SharedCapturer<'a> {
   }
 
   fn check_buffer(&self) -> Result<()> {
-    // TODO: is this needed to be checked every time?
     if self.buffer_size < self.dxgi_outdupl_desc().calc_buffer_size() {
       Err(Error::new("Invalid buffer length"))
     } else {
@@ -192,13 +185,6 @@ impl<'a> Capturer for SharedCapturer<'a> {
 
   fn pointer_shape_buffer(&self) -> &[u8] {
     &self.pointer_shape_buffer[..self.pointer_shape_buffer_size]
-  }
-
-  fn pointer_shape_updated(&self) -> bool {
-    self.pointer_shape_buffer_size != self.last_pointer_shape_buffer_size || {
-      let len = self.pointer_shape_buffer_size;
-      self.pointer_shape_buffer[..len] != self.last_pointer_shape_buffer[..len]
-    }
   }
 
   fn capture(&mut self) -> Result<DXGI_OUTDUPL_FRAME_INFO> {
@@ -226,23 +212,12 @@ impl<'a> Capturer for SharedCapturer<'a> {
       self.buffer_size,
       &self.texture,
       &self.texture_desc,
-      &mut self.last_pointer_shape_buffer, // IMPORTANT: write to last pointer shape buffer
+      &mut self.pointer_shape_buffer,
     )?;
 
     if pointer_shape_info.is_some() {
       // record the pointer shape buffer size
-      // IMPORTANT: change last pointer shape buffer size
-      self.last_pointer_shape_buffer_size = frame_info.PointerShapeBufferSize as usize;
-
-      // swap the pointer shape buffer
-      std::mem::swap(
-        &mut self.pointer_shape_buffer,
-        &mut self.last_pointer_shape_buffer,
-      );
-      std::mem::swap(
-        &mut self.pointer_shape_buffer_size,
-        &mut self.last_pointer_shape_buffer_size,
-      );
+      self.pointer_shape_buffer_size = frame_info.PointerShapeBufferSize as usize;
     }
 
     Ok((frame_info, pointer_shape_info))